Modularización del servicio de autorización
